+++ src/petals/server/reachability.py
@@ -1,16 +1,29 @@
+import asyncio
 import math
+import threading
 import time
+from contextlib import asynccontextmanager
+from functools import partial
+from typing import Optional
 
 import requests
-from hivemind.utils.logging import get_logger
+from hivemind.dht import DHT, DHTNode
+from hivemind.moe.client.remote_expert_worker import RemoteExpertWorker
+from hivemind.p2p import P2P, P2PContext, PeerID, ServicerBase
+from hivemind.proto import dht_pb2
+from hivemind.utils import get_logger
 
-logger = get_logger(__file__)
+from petals.constants import VALIDATOR_URL
+
+logger = get_logger(__name__)
+PROBE_P2P_ARGS = dict(dht_mode="client", use_relay=False, auto_nat=False, nat_port_map=False, no_listen=True)
 
 
-def check_reachability(peer_id, wait_time: float = 7 * 60, retry_delay: float = 15) -> None:
+def validate_reachability(peer_id, wait_time: float = 7 * 60, retry_delay: float = 15) -> None:
+    """verify that your peer is reachable from a (centralized) validator, whether directly or through a relay"""
     for attempt_no in range(math.floor(wait_time / retry_delay) + 1):
         try:
-            r = requests.get(f"http://health.petals.ml/api/v1/is_reachable/{peer_id}", timeout=10)
+            r = requests.get(f"{VALIDATOR_URL}/api/v1/is_reachable/{peer_id}", timeout=10)
             r.raise_for_status()
             response = r.json()
 
@@ -37,3 +50,77 @@ def check_reachability(peer_id, wait_time: float = 7 * 60, retry_delay: float = 
         f"        python -m petals.cli.run_server ... --public_ip {response['your_ip']} --port 31337\n"
         f"    4. If it does not help, ask for help in our Discord: https://discord.gg/Wuk8BnrEPH\n"
     )
+
+
+def check_direct_reachability(max_peers: int = 5, threshold: float = 0.5, **kwargs) -> Optional[bool]:
+    """test if your peer is accessible by others in the swarm with the specified network options in **kwargs"""
+
+    async def _check_direct_reachability():
+        dht_tester = await DHTNode.create(client_mode=True, **kwargs)
+        protocol = ReachabilityProtocol(dht_tester.protocol.p2p)
+        async with protocol.serve():
+            successes = requests = 0
+            for remote_peer in list(dht_tester.protocol.routing_table.peer_id_to_uid.keys()):
+                probe_available = await protocol.call_check(remote_peer=remote_peer, check_peer=dht_tester.peer_id)
+                if probe_available is None:
+                    continue  # remote peer failed to check probe
+                successes += probe_available
+                requests += 1
+                if requests >= max_peers:
+                    break
+        await dht_tester.shutdown()
+        logger.debug(f"Reachability: observed {successes} successes out of {requests} requests")
+        return (successes / requests) >= threshold if requests > 0 else None
+
+    return RemoteExpertWorker.run_coroutine(_check_direct_reachability())
+
+
+class ReachabilityProtocol(ServicerBase):
+    """Mini protocol to test if a locally running peer is accessible by other devices in the swarm"""
+
+    def __init__(self, p2p: P2P, *, probe: Optional[P2P] = None, wait_timeout: float = 5.0):
+        probe = probe if probe is not None else p2p
+        self.p2p, self.probe, self.wait_timeout = p2p, probe, wait_timeout
+
+    async def call_check(self, remote_peer: PeerID, *, check_peer: PeerID) -> Optional[bool]:
+        """return True if remote_peer can reach check_peer, False if cannot, None means remote_peer did not respond"""
+        try:
+            request = dht_pb2.PingRequest(peer=dht_pb2.NodeInfo(node_id=check_peer.to_bytes()))
+            timeout = self.wait_timeout if check_peer == remote_peer else self.wait_timeout * 2
+            response = await self.get_stub(self.probe, remote_peer).rpc_check(request, timeout=timeout)
+            return response.available
+        except Exception as e:
+            logger.debug(f"requested {remote_peer} to check {check_peer}, but got {repr(e)}", exc_info=True)
+
+    async def rpc_check(self, request: dht_pb2.PingRequest, context: P2PContext) -> dht_pb2.PingResponse:
+        """Another peer wants us to help it check reachability"""
+        response = dht_pb2.PingResponse(available=True)
+        check_peer = PeerID(request.peer.node_id)
+        if check_peer != context.local_id:  # remote peer wants us to check someone other than ourselves
+            response.available = await self.call_check(check_peer, check_peer=check_peer) is True
+        return response
+
+    @asynccontextmanager
+    async def serve(self):
+        try:
+            await self.add_p2p_handlers(self.p2p)
+            yield self
+        finally:
+            await self.remove_p2p_handlers(self.p2p)
+
+    @classmethod
+    def attach_to_dht(cls, dht: DHT, **kwargs):
+        async def _serve_with_probe():
+            protocol = cls(p2p=await dht.replicate_p2p(), **kwargs)
+            initial_peers = list(map(str, await protocol.p2p.get_visible_maddrs(latest=True)))
+            for info in await protocol.p2p.list_peers():
+                initial_peers.extend((f"{addr}/p2p/{info.peer_id.to_base58()}" for addr in info.addrs))
+            protocol.probe = await P2P.create(initial_peers, **PROBE_P2P_ARGS)
+            try:
+                async with protocol.serve():
+                    while dht.is_alive():
+                        await asyncio.sleep(protocol.wait_timeout)  # not awaiting join because it freezes dht.shutdown
+            finally:
+                await protocol.probe.shutdown()
+
+        threading.Thread(target=partial(asyncio.run, _serve_with_probe()), daemon=True).start()
